microservices/auth/src/utils/hashing_helper.py
from passlib.context import CryptContext
from datetime import datetime, timedelta
from typing import Union
from ..config.config import settings
from jose import jwt


password_context = CryptContext(schemes=["bcrypt"], deprecated="auto")


def get_hashed_password(password: str) -> str:
    return password_context.hash(password)

def verify_password(password: str, hashed_pass: str) -> bool:
    return password_context.verify(password, hashed_pass)

def create_access_token(data: dict, expires_delta: Union[timedelta, None] = None):
    access_token_expires = timedelta(minutes=settings.access_token_expire_minutes)

    to_encode = data.copy()

    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=5)
    to_encode.update({"exp": expire})

    return jwt.encode(to_encode, settings.secret_key, algorithm=settings.algorithm)
