microservices/auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = "Authentication and Authorization microservice"
authors = ["Dilip Kumar Sharma <diliplakshya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.89.1"}
sqlalchemy = "^2.0.0"
pymysql = "^1.0.2"
python-multipart = "^0.0.5"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
